Make web3 contract less accessible; and wrappers be the first choice
